Base training

+++ src/DoodleClassifier/DoodleClassifier/Dataset/Categories.cs
@@ -36,7 +36,7 @@ namespace DoodleClassifier
 			}
 		}
 
-		public static int Count => categories.Count;
+		public static uint Count => (uint)categories.Count;
 		public static IReadOnlyList<string> Enumerate() => categories;
 
 		public static bool IsValid(string name) => oneHots.ContainsKey(name);
